# openapi.yaml
openapi: 3.0.3
info:
  title: ai-model-arena API
  version: 1.0.0
  description: |
    Public REST API for ai-model-arena — multi-model agentic coding arena.
    All endpoints are versioned under `/api/v1/` and protected by API-key
    auth (`X-API-Key` header) with per-key rate limiting. Interactive docs are
    served at `/api/docs`. Run-management endpoints used by the dashboard SPA
    are also available under `/api/*` via JWT auth.
  contact:
    name: ai-model-arena
servers:
  - url: http://localhost:4000/api/v1
    description: Local dashboard server
security:
  - ApiKeyAuth: []
...
paths:
...
  /runs:
    get:
      tags: [runs]
      summary: List runs (newest first)
      responses: { '200': { description: Run list } }
    post:
      tags: [runs]
      summary: Trigger a new run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenario, models]
              properties:
                scenario: { type: string }
                models: { type: array, items: { type: string } }
      responses: { '202': { description: Run started } }
...
  /anomalies:
    get:
      tags: [anomalies]
      summary: List detected anomalies
      parameters:
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/AnomalyType' }
        - in: query
          name: severity
          schema: { $ref: '#/components/schemas/AnomalySeverity' }
        - in: query
          name: resolved
          schema: { type: boolean }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses: { '200': { description: Anomaly list } }
...
  /webhooks:
    get:
      tags: [webhooks]
      summary: List registered webhooks
      responses: { '200': { description: Webhook list } }
    post:
      tags: [webhooks]
      summary: Register a webhook URL for events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items:
                    type: string
                    enum: [run_completed, anomaly_detected, budget_exceeded]
                secret: { type: string }
      responses: { '201': { description: Webhook registered } }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from configs/api-keys.yaml. Rate-limited per key.
...
  schemas:
...
    AnomalyType:
      type: string
      enum: [latency, loop, token_spike, cost_spike, error_rate, silent_failure]
    AnomalySeverity:
      type: string
      enum: [low, medium, high, critical]